modified post sql queries

+++ 5-logic/postLogic.ts
@@ -2,15 +2,37 @@ import { OkPacket } from "mysql2";
 import { execute } from "../2-utils/dal";
 import { PostModel } from "../4-models/PostModel";
 
-export async function getPostsByUser(userId: number): Promise<PostModel[]> {
-  const query = `SELECT * FROM posts WHERE userId = ?`;
-  const [res] = await execute<PostModel[]>(query, [userId]);
+export async function getPostsByUser(
+  sessionUserId: number,
+  postUserId: number
+): Promise<PostModel[]> {
+  const params = [sessionUserId, postUserId];
+  // subquery userId = sessionUserId, query userId = postUserId
+  const query = `SELECT p.*, count(l.postId) AS likes, 
+  CASE WHEN EXISTS (SELECT 1 FROM likes WHERE postId = p.postImg AND userId = ?)
+       THEN true END AS isLiked
+FROM posts p
+LEFT JOIN likes l ON p.postImg = l.postId
+WHERE p.userId = ?
+GROUP BY p.postImg`;
+  const [res] = await execute<PostModel[]>(query, params);
   return res;
 }
 
-export async function getPostByPostId(postId: number): Promise<PostModel> {
-  const query = `SELECT * FROM posts WHERE postImg = ?`;
-  const [res] = await execute<PostModel[]>(query, [postId]);
+export async function getPostByPostId(
+  userId: number,
+  postId: number
+): Promise<PostModel> {
+  const params = [userId, postId];
+
+  const query = `SELECT p.*, count(l.postId) AS likes, 
+  CASE WHEN EXISTS (SELECT 1 FROM likes WHERE postId = p.postImg AND userId = ?)
+       THEN true END AS isLiked
+FROM posts p
+LEFT JOIN likes l ON p.postImg = l.postId
+WHERE p.postImg = ?
+GROUP BY p.postImg`;
+  const [res] = await execute<PostModel[]>(query, params);
   return res[0];
 }
 
@@ -24,8 +46,11 @@ export async function createPost(post: PostModel): Promise<OkPacket> {
 
 export async function updatePost() {} //check if neended
 
-export async function deletePost(postId: number): Promise<OkPacket> {
-  const query = `DELETE FROM posts WHERE postImg = ?`;
-  const [res] = await execute<OkPacket>(query, [postId]);
+export async function deletePost(
+  postId: number,
+  userId: number
+): Promise<OkPacket> {
+  const query = `DELETE FROM posts WHERE postImg = ? AND userId = ?`;
+  const [res] = await execute<OkPacket>(query, [postId, userId]);
   return res;
 }
